+++ src/main/kotlin/uk/co/nickthecoder/tickle/action/movement/Drag.kt
@@ -1,17 +1,17 @@
 package uk.co.nickthecoder.tickle.action.movement
 
-import org.joml.Vector2f
+import org.joml.Vector2d
 import uk.co.nickthecoder.tickle.action.Action
 
 class Drag(
-        val velocity: Vector2f,
-        var drag: Float)
+        val velocity: Vector2d,
+        var drag: Double)
 
     : Action {
 
 
     override fun act(): Boolean {
-        if (drag != 0f) {
+        if (drag != 0.0) {
             velocity.mul(1 - drag)
         }
         return false
